Remove review usefulness normalisation by age; resplit data
Previously the usefulness votes were normalised by age i.e. each
review's usefulness votes divided by the age of the review at the time
of data collection. However, data collection date was potentially
erroneously assumed to be constant and on the date of the latest review.

The reality is that there is no data collection date information in the
dataset, so the best we can do is just not normalise by review age.
Attempting to do so may introduce bias.